remove logger

+++ GalileoFront/app/src/main/java/com/example/galileomastermindboilerplate/SatelliteWidgetEntryData.java
@@ -16,7 +16,7 @@ public class SatelliteWidgetEntryData {
     public double AccumulatedDeltaRangeMeters = -1;
     public double AccumulatedDeltaRangeUncertaintyMeters = -1;
     public double CarrierFrequencyHz = -1;
-    public double CarrierCycles = -1;
+    public long CarrierCycles = -1;
     public double CarrierPhase = -1;
     public double FullInterSignalBiasNanos = -1;
     public double FullInterSignalBiasUncertaintyNanos = -1;
